refactor: explorer as package and external component
