+++ frontend/package.json
@@ -45,7 +45,7 @@
   },
   "dependencies": {
     "@edifice-ui/icons": "develop",
-    "@edifice-ui/react": "develop",
+    "@edifice-ui/react": "file:edifice-ui-react-1.3.1.tgz",
     "@edifice-ui/explorer": "file:edifice-ui-explorer-1.0.0.tgz",
     "@edifice-wisemapping/editor": "0.7.8",
     "@react-spring/web": "9.7.3",
